+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+- The [](./providers.net.launchpad.git.md) provider for [Launchpad](https://launchpad.net) has been added.
 - The `unicorn.util.substituteInPlace` function has been added.
 
 ## v1.3.1